feat: allow backup to persist and restore streaming id

+++ crates/enterprise/unleash-edge-enterprise-integration-tests/src/lib.rs
@@ -163,11 +163,11 @@ mod tests {
             unimplemented!()
         }
 
-        async fn save_last_event_ids(&self, _event_id: HashMap<String, u64>) -> EdgeResult<()> {
+        async fn save_last_event_ids(&self, _event_id: HashMap<String, u32>) -> EdgeResult<()> {
             unimplemented!()
         }
 
-        async fn load_last_event_ids(&self) -> EdgeResult<HashMap<String, u64>> {
+        async fn load_last_event_ids(&self) -> EdgeResult<HashMap<String, u32>> {
             unimplemented!()
         }
     }
